% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diminui.R
\name{diminui}
\alias{diminui}
\title{Função diminui}
\usage{
diminui(a, b)
}
\description{
Esta é uma função simples para diminuir dois números
}
\details{
@param a Primeiro número
 @param b Segundo número

  @examples
  diminui(3,2)
  diminui(5,3)
}
